Add RenderManager2D
This is a nice abstraction layer/interface to the new vulkan renderer. Hopefully I can hook textures in the next commit and then add cross-API imgui support. After that, I need to rewrite the OpenGL side of things to be more friendly to the Vulkan style renderer I have now.